Log: Add option to set LevelFilter at compile time
If the Environment variable HERMIT_LOG_LEVEL_FILTER is set at compile time to a string matching the Name of a Levelfilter enum value, then that value is used for the Levelfilter.
If the Environment variable is not set, or the name doesn't match, then LevelFilter::Info is used by default, which is the same as it was before.