combine master and devel branch together
To simplify the development, the feature "with_submodule" is introduced.
If this is set, a submodule with the kernel is expected and use this
submodule to build the library operating system. Otherwise,
it use the source code from crates.io.
Per default, the code from crates.io is used. However, the examples of this
repository use the kernel as submodule.