Changed syscall for new network driver
This commit changes the syscall interface slightly for the new network
driver. Now upon the receiption of a packet, the driver does forward a
reference to the packets memory as also a handle, which controls this
memory area inside the kernel-space driver part. The handle is then
returned to the network driver upon consumption. This allows to reduce
the amount of necessary copies by one.