remove deadlock in the network stack
To avoid deadlocks,the  mutex and the device checks are revised.
In principle, the network thread doesn't block, if another thread
is already checking the device.

In addition, the send and receive operations are optimized to stay longer
in polling mode and assigns the network handler to the core, where
the network thread is running. This should reduce the number of
wakeup interrupts.